requirements.txt
django-bootstrap3
selectable